+++ exonum-java-binding-fakes/src/main/java/com/exonum/binding/fakes/services/service/PutValueTransaction.java
@@ -22,8 +22,8 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.common.hash.Hashing;
 import com.exonum.binding.storage.indices.ProofMapIndexProxy;
-import com.exonum.binding.transaction.AbstractTransaction;
 import com.exonum.binding.transaction.RawTransaction;
+import com.exonum.binding.transaction.Transaction;
 import com.exonum.binding.transaction.TransactionContext;
 import java.nio.ByteBuffer;
 import java.nio.CharBuffer;
@@ -33,7 +33,7 @@ import java.nio.charset.CharsetDecoder;
 import java.nio.charset.CodingErrorAction;
 import java.nio.charset.StandardCharsets;
 
-final class PutValueTransaction extends AbstractTransaction {
+final class PutValueTransaction implements Transaction {
 
   @SuppressWarnings("WeakerAccess")  // Might be used in native tests.
   static final short ID = 0x1234;
@@ -47,7 +47,7 @@ final class PutValueTransaction extends AbstractTransaction {
     checkArgument(rawTransaction.getServiceId() == TestService.ID);
     checkArgument(rawTransaction.getTransactionId() == PutValueTransaction.ID);
     String value = getValue(rawTransaction);
-    return new PutValueTransaction(rawTransaction, value, checkNotNull(schemaFactory));
+    return new PutValueTransaction(value, checkNotNull(schemaFactory));
   }
 
   private static String getValue(RawTransaction rawTransaction) {
@@ -67,9 +67,8 @@ final class PutValueTransaction extends AbstractTransaction {
         .onUnmappableCharacter(CodingErrorAction.REPLACE);
   }
 
-  private PutValueTransaction(RawTransaction rawTransaction, String value,
+  private PutValueTransaction(String value,
       SchemaFactory<TestSchema> schemaFactory) {
-    super(rawTransaction);
     this.value = value;
     this.schemaFactory = schemaFactory;
   }