+++ exonum-java-binding-qa-service/src/main/java/com/exonum/binding/qaservice/transactions/CreateCounterTx.java
@@ -66,11 +66,6 @@ public final class CreateCounterTx implements Transaction {
   }
 
   @Override
-  public HashCode hash() {
-    return converter().toRawTransaction(this).hash();
-  }
-
-  @Override
   public boolean equals(Object o) {
     if (this == o) {
       return true;